wip: add a module system
the evaluator tests still need to be moved into their respective modules, and made only to depend on what they need.

this also removes syscalls, since a more sandboxed approach is going to be preferred moving forward.

this doesn't currently work with the scanner since functions are stored on the program.