Add testing utilities (#22)
As of #21, expressions are now split into two structs. This makes
testing and asserting very tedious as one would need to know the exact
span of each expression in order to assert that it is equal to an
expression on the stack.

Though it is useful to test that the parser is properly measuring spans,
it is not useful for higher-level tests that only care about the values.

Instead, I have implemented a `ExprSimple` struct that resembles the
non-split version of an expression. This will only be used by the
previously mentioned "higher-level" tests. Along with this struct, I
have also added a few utility functions for converting expressions and a
vec of expressions to the simpler version.

Lastly, I have also reimplemented all tests except for the full parser
suite.

---------

Co-authored-by: leonskidev <leonskidev@pm.me>