Prevent recursion for the `recur` keyword (#33)
When you use the `recur` keyword, it tells the engine to rerun the outer
function. Instead of creating recursion in the language, Stack would
simply keep the scope around and rerun the code within the function.

However, in Rust, the engine was still recursing under the hood through
the `Engine::call_fn` method. To fix this problem, I refactored the code
to utilize a loop which checks if recursion is requested reruns the
`Engine::call_fn` method through an outer method, preventing recursion
on the Rust level.