Add debug hook (#38)
Adds a debug hook function so embedders of Stack can provide an
appropriate debug print if it makes sense to do so. This allows users of
Stack to have some way to get debug output, even if an alternative is
not available; especially useful once `print` and `pretty` are no-longer
intrinsics and become part of the standard library.